requirements.txt
opencv-python
torch
numpy
PyYAML
torchvision
ipykernel
matplotlib